AI: moved logic to backend
