+++ package.json
@@ -1,0 +1,23 @@
+{
+  "name": "sports-bet",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "echo index.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/thealinivan/sports-bet.git"
+  },
+  "author": "alin ivan",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/thealinivan/sports-bet/issues"
+  },
+  "homepage": "https://github.com/thealinivan/sports-bet#readme",
+  "dependencies": {
+    "express": "^4.17.1"
+  }
+}
